connector: load loopback models from filesystem
Add a new internal property `ModelDefinition.readonly` to signal whether
the definition can be edited by users.

Modify connector to search `node_modules/loopback` for model
definitions.

Fix `ConfigFile` methods to correctly support subclassing.

+++ models/model-definition.js
@@ -103,6 +103,9 @@ ModelDefinition.toFilename = function(name) {
 
 ModelDefinition.removeById = function(id, cb) {
   this.findById(id, function(err, modelDef) {
+    if (modelDef.readonly) {
+      return cb(new Error('Cannot remove readonly model ' + id));
+    }
     var p = ModelDefinition.getPath(modelDef.facetName, modelDef);
     var file = new ConfigFile({path: p});
     file.remove(cb);
